Fix page layout
Use `data-theme="wide"` width custom column widths.

Also reorganizes sections and fixes `z-index` of `<nav>` & `<button is="share-button">`.
