+++ _data/nav.json
@@ -3,10 +3,6 @@
   "href": "/#rooms",
   "icon": "hotel"
 }, {
-  "label": "Map",
-  "href": "/#map",
-  "icon": "map"
-}, {
   "label": "Events",
   "href": "/#events",
   "icon": "event"
@@ -14,4 +10,8 @@
   "label": "Contact",
   "href": "/#contact",
   "icon": "comment-discussion"
+}, {
+  "label": "Map",
+  "href": "/#map",
+  "icon": "map"
 }]
