+++ _includes/sections/rooms.html
@@ -39,7 +39,7 @@
 	</div>
 	<dialog id="accessibility" class="animation-speed-normal animation-ease-in-out fadeIn">
 		<header class="sticky top clearfix requires-js">
-			<a href="{{ page.url }}#" role="button" class="btn btn-primary float-right" title="close">
+			<a href="{{ page.url }}#" role="button" class="btn btn-transparent float-right" title="close">
 				{% include icon.html icon="x" %}
 			</a>
 		</header>
