+++ _includes/share-button.html
@@ -1,3 +1,3 @@
-<button type="button" is="share-button" class="round shadow fixed z-top" aria-label="Share" text="{{ include.text | default: page.description }}"{% if include.url %} url="{{ include.url }}"{% endif %} hidden="">
+<button type="button" is="share-button" class="round shadow fixed z-4" aria-label="Share" text="{{ include.text | default: page.description }}"{% if include.url %} url="{{ include.url }}"{% endif %} hidden="">
 	{% include icon.html icon="public-share" %}
 </button>
