+++ css/index.css
@@ -12,6 +12,10 @@
 @import url("./vars.css");
 @import url("./layout.css");
 
+:root[data-layout="wide"] > body.grid {
+	grid-template-columns: auto minmax(320px, 1200px) auto;
+}
+
 input[hidden] + label[for].btn-toggle {
 	background-color: var(--button-disabled-background);
 	margin: 0.3rem;
@@ -96,7 +100,11 @@ leaflet-map:not(:defined) [slot="icon"] {
 	margin-bottom: 0.6em;
 }
 
-.btn-wide {
+.btn.btn-wide {
 	width: 80%;
 	margin-left: 10%;
+}
+
+.btn.btn-transparent {
+	background-color: transparent;
 }
