+++ index.html
@@ -4,6 +4,7 @@ title: Home
 pinned: true
 referrer: no-referrer
 description: Family owner motel in Kernville, CA since 1955
+layout: wide
 keywords:
 - Piazza
 - Pine Cone
@@ -14,13 +15,7 @@ keywords:
 - kernville
 - lake isabella
 ---
-{% include nav.html %}
-{% include header.html %}
-<main id="main">
-	{% include sections/rooms.html %}
-	{% include sections/map.html %}
-	{% include sections/events.html %}
-	{% include sections/contact.html %}
-</main>
-{% include footer.html %}
-{% include share-button.html %}
+{% include sections/rooms.html %}
+{% include sections/events.html %}
+{% include sections/contact.html %}
+{% include sections/map.html %}
